Remove warning on Solaris
git-svn-id: http://svn.osgeo.org/postgis/trunk@5011 b70326c6-7e19-0410-871a-916f4a2858ee

+++ postgis/lwgeom_chip.c
@@ -172,7 +172,7 @@ Datum CHIP_in(PG_FUNCTION_ARGS)
 	{
 		if (result->size != (sizeof(CHIP) - sizeof(void*) + datum_size * result->width*result->height) )
 		{
-			elog(ERROR,"CHIP_in parser - bad data (actual size [%d] != computed size [%ld])!", result->size, (sizeof(CHIP)-sizeof(void*) + datum_size * result->width*result->height) );
+			elog(ERROR,"CHIP_in parser - bad data (actual size [%d] != computed size [%ld])!", result->size, (long int)(sizeof(CHIP)-sizeof(void*) + datum_size * result->width*result->height) );
 			PG_RETURN_NULL();
 		}
 	}
